test: useFunnel 테스트 [#5]
